securespot/requirements.txt
Pillow==10.4.0
fastapi==0.112.0
uvicorn==0.30.5
pydantic==2.8.2
numpy
google-generativeai
ultralytics
googlemaps
dnspython
langchain_google_genai
pymongo
torch
langchain-community
passlib
python-jose
motor
pydantic[email]
python-multipart
markdown



